requirements.txt
strands-agents == 1.7.1
boto3 == 1.40.25
python-dotenv == 1.1.1
mcp == 1.13.1
tenacity == 9.1.2
PyGithub == 2.8.1

rich == 14.1.0
prompt_toolkit == 3.0.52
